requirements.txt
python-decouple==3.8
fastapi==0.115.5
uvicorn==0.32.0
confluent_kafka==2.6.0
requests==2.32.3
openmeteo-requests
requests-cache
retry-requests
pandas
pydantic